+++ shared/src/main/scala/doodle/random.scala
@@ -4,7 +4,7 @@ import cats.{Comonad,Monad}
 import cats.free.Free
 import scala.util.{Random => Rng}
 
-package object random {
+object random {
   type Random[A] = Free[RandomOp,A]
 
   sealed abstract class RandomOp[A]
@@ -13,7 +13,7 @@ package object random {
     final case object RInt extends RandomOp[Int]
     final case class Natural(upperLimit: Int) extends RandomOp[Int]
     final case object RDouble extends RandomOp[Double]
-    final case object Gaussian extends RandomOp[Double]
+    final case object Normal extends RandomOp[Double]
   }
 
   implicit val randomInstance: Monad[Random] = Free.freeMonad
@@ -30,9 +30,40 @@ package object random {
           case RInt => rng.nextInt()
           case Natural(u) => rng.nextInt(u)
           case RDouble => rng.nextDouble()
-          case Gaussian => rng.nextGaussian()
+          case Normal => rng.nextGaussian()
         }
       override def map[A, B](fa: RandomOp[A])(f: (A) ⇒ B): RandomOp[B] =
         Always(f(extract(fa)))
     }
+
+  object Random {
+    import RandomOp._
+
+    def always[A](in: A): Random[A] =
+      Free.pure(in)
+
+    def int: Random[Int] =
+      Free.liftF[RandomOp,Int](RInt)
+
+    def natural(upperLimit: Int): Random[Int] =
+      Free.liftF[RandomOp,Int](Natural(upperLimit))
+
+    def double: Random[Double] =
+      Free.liftF[RandomOp,Double](RDouble)
+
+    def oneOf[A](elts: A*): Random[A] = {
+      val length = elts.length
+      Random.natural(length) map (idx => elts(idx))
+    }
+
+    def discrete[A](elts: (A, Double)*): Random[A] =
+      ???
+
+    def normal: Random[Double] =
+      Free.liftF[RandomOp,Double](Normal)
+
+    def normal(mean: Double, stdDev: Double): Random[Double] =
+      normal map (x => (stdDev * x) + mean)
+  }
+
 }
